% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause_map.R
\name{cause_map}
\alias{cause_map}
\title{Deriving Broad Cause of Death from CCVA Outputs}
\usage{
cause_map(df, age_group)
}
\arguments{
\item{df}{Outputs from \code{codEAVA()} in \code{EAVA} for EAVA, and \code{codeVA()} and \code{prepCalibration()} in \code{openVA} for InSilicoVA and InterVA}

\item{age_group}{Character. Indicates age group.

\code{"neonate"} for deaths with 0-27 days of birth, and \code{"child"} for 1-59 months of birth.}
}
\value{
Matrix. Rows are individuals. Columns are broad causes.
This is a binary matrix (entries 0 or 1) with 1 indicating the broad cause of death for the individual.
}
\description{
Takes individual-level cause of deaths (output from CCVA algorithms) as input, and maps them to pre-defined broad causes.
}
\examples{

## Publicly Available Cause-of-Death (COD) Data from COMSA–Mozambique
comsamoz_CCVAoutput$neonate$eava # output from EAVA algorithm for age group "neonate"
head(comsamoz_CCVAoutput$neonate$eava)  # specific COD for the first 6 deaths

## broad cause mapping
mapped_broad_cause = cause_map(df = comsamoz_CCVAoutput$neonate$eava, age_group = "neonate")
head(mapped_broad_cause)  # broad COD for the first 6 deaths

}
