% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comsamoz_CCVAoutput.R
\docType{data}
\name{comsamoz_CCVAoutput}
\alias{comsamoz_CCVAoutput}
\title{CCVA Outputs for Publicly Available Verbal Autopsy (VA) Data from COMSA–Mozambique}
\format{
List.
\describe{
\item{neonate}{List. Outputs of EAVA, InSilicoVA, and InterVA for \code{"neonate"} (0-27 days)}
\item{child}{List. Outputs of EAVA, InSilicoVA, and InterVA for \code{"child"} (1-59 months)}
\item{version}{Character. Date stamp (yyyymmdd) for version control. Only for package maintainers.}
}
}
\usage{
comsamoz_CCVAoutput
}
\description{
This contains outputs of CCVA algorithms \href{https://www.jogh.org/documents/issue201601/jogh-06-010601.pdf}{EAVA}, \href{https://www.tandfonline.com/doi/abs/10.1080/01621459.2016.1152191}{InSilicoVA}, and InterVA (\doi{10.3402/gha.v5i0.19281}) when applied to publicly available verbal autopsy (VA) data collected in the Countrywide Mortality Surveillance for Action project in Mozambique (\href{https://comsamozambique.org/}{COMSA-Mozambique}).
}
\details{
Outputs for EAVA are obtained using the \href{https://CRAN.R-project.org/package=EAVA}{EAVA} package, while outputs for InSilicoVA and InterVA are produced using the \href{https://CRAN.R-project.org/package=openVA}{openVA} package.

For example, \code{comsamoz_CCVAoutput$neonate$eava} contains output from the EAVA algorithm for "neonate".
}
\references{
Pramanik, S, et al. (2026)
Country-Specific Estimates of Misclassification Rates of Computer-Coded Verbal Autopsy Algorithms
BMJ Global Health
\doi{10.1136/bmjgh-2025-021747}

Pramanik, S, et al. (2025)
Modeling structure and country-specific heterogeneity in misclassification matrices of verbal autopsy-based cause of death classifiers
Annals of Applied Statistics
\href{https://projecteuclid.org/journals/annals-of-applied-statistics/volume-19/issue-2/Modeling-structure-and-country-specific-heterogeneity-in-misclassification-matrices-of/10.1214/24-AOAS2006.short}{Link}

Wilson E, et al. (2025)
EAVA: Deterministic Verbal Autopsy Coding with Expert Algorithm Verbal Autopsy
\href{https://CRAN.R-project.org/package=EAVA}{Link}

Zehang Richard Li, et al. (2024)
openVA: Automated Method for Verbal Autopsy
R package version 1.1.2.
\href{https://CRAN.R-project.org/package=openVA}{Link}

Countrywide Mortality Surveillance for Action in Mozambique (COMSA-Mozambique).
\href{https://comsamozambique.org/}{Link}

Macicame, I, et al. (2023)
Countrywide Mortality Surveillance for Action in Mozambique: Results from a National Sample-Based Vital Statistics System for Mortality and Cause of Death
American Journal of Tropical Medicine and Hygiene
\doi{10.4269/ajtmh.22-0367}

Zehang Richard Li, et al. (2023)
The openVA Toolkit for Verbal Autopsies
The R Journal \href{https://journal.r-project.org/articles/RJ-2023-020/}{Link}

Kalter, H., et al. (2016)
Validating hierarchical verbal autopsy expert algorithms in a large data set with known causes of death.
Journal of Glob Health
\href{https://www.jogh.org/documents/issue201601/jogh-06-010601.pdf}{Link}

McCormick, Tyler H., et al. (2016)
Probabilistic Cause-of-Death Assignment Using Verbal Autopsies
Journal of the American Statistical Association
\href{https://www.tandfonline.com/doi/abs/10.1080/01621459.2016.1152191}{Link}

Byass, Peter, et al. (2012)
Strengthening standardised interpretation of verbal autopsy data: the new InterVA-4 tool
Global Health Action
\doi{10.3402/gha.v5i0.19281}
}
\keyword{datasets}
