% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_vacalib_prior.R
\name{modular_vacalib_prior}
\alias{modular_vacalib_prior}
\title{Modular VA-Calibration using Dirichlet Prior on Misclassification Matrix}
\usage{
modular_vacalib_prior(
  va_unlabeled,
  Mmat_calib,
  studycause_map,
  donotcalib,
  donotcalib_type,
  nocalib.threshold,
  path_correction,
  ensemble,
  pshrink_strength,
  nMCMC,
  nBurn,
  nThin,
  nChain,
  nCore,
  adapt_delta_stan,
  refresh_stan,
  seed,
  verbose,
  input_vacalib
)
}
\arguments{
\item{va_unlabeled}{Same as \code{va_unlabeled} in \code{vacalibration()}}

\item{Mmat_calib}{Same as \code{missmat} in \code{vacalibration()}}

\item{studycause_map}{Same as \code{studycause_map} in \code{vacalibration()}}

\item{donotcalib}{Same as \code{donotcalib} in \code{vacalibration()}}

\item{donotcalib_type}{Same as \code{donotcalib_type} in \code{vacalibration()}}

\item{nocalib.threshold}{Same as \code{nocalib.threshold} in \code{vacalibration()}}

\item{path_correction}{Same as \code{path_correction} in \code{vacalibration()}}

\item{ensemble}{Same as \code{ensemble} in \code{vacalibration()}}

\item{pshrink_strength}{Same as \code{pshrink_strength} in \code{vacalibration()}}

\item{nMCMC, nBurn, nThin}{Same as \code{nMCMC}, \code{nBurn}, and \code{nThin} in \code{vacalibration()}}

\item{nChain}{Same as \code{nChain} in \code{vacalibration()}}

\item{nCore}{Same as \code{nCore} in \code{vacalibration()}}

\item{adapt_delta_stan}{Same as \code{adapt_delta_stan} in \code{vacalibration()}}

\item{refresh_stan}{Same as \code{refresh_stan} in \code{vacalibration()}}

\item{seed}{Same as \code{seed} in \code{vacalibration()}}

\item{verbose}{Same as \code{verbose} in \code{vacalibration()}}

\item{input_vacalib}{List of inputs in \code{vacalibration()}}
}
\value{
Similar to the list returned in \code{vacalibration()}
}
\description{
This is a utility function. Please use \link{vacalibration}.
}
