% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vacalib.R
\name{plot_vacalib}
\alias{plot_vacalib}
\title{Summary Plots of VA-Calibration}
\usage{
plot_vacalib(vacalib_fit, toplot = "both")
}
\arguments{
\item{vacalib_fit}{Fitted object from \code{vacalibration()}}

\item{toplot}{Character. What to plot.

When \code{toplot="missmat"} and \code{missmat_type="fixed"}, it plots the fixed misclassification matrix used in calibration. When \code{missmat_type} equals \code{"fixed"} or \code{"samples"}, it plots the average misclassification matrix.

When \code{toplot="csmf"}, it compares uncalibrated and calibrated estimates of cause-specific mortality fractions (CSMFs).

When \code{"both"}, it plots both the misclassification matrix and estimates of CSMFs.}
}
\value{
It returns a plot comparing misclassification matrix used in calibration, and uncalibrated and calibrated estimates of cause-specific mortality fractions (CSMFs).
}
\description{
Given a VA-Calibration fit using \link{vacalibration}, this function plots misclassification matrix used in VA-Calibration, and compares uncalibrated and calibrated estimates of cause-specific mortality fractions (CSMFs).
}
\examples{

\donttest{
######### COMSA-Mozambique VA-COD data #########
data(comsamoz_CCVAoutput)

######### Algorithm-Specific Calibration #########

# EAVA
vacalib_out_eava = vacalibration(va_data = comsamoz_CCVAoutput$neonate[1],
                                 age_group = "neonate", country = "Mozambique",
                                 saveoutput = FALSE)
print(vacalib_out_eava$input$missmat_type)
print(vacalib_out_eava$input)
print(names(vacalib_out_eava$input))

# summary plot
plot_vacalib(vacalib_fit = vacalib_out_eava, toplot = "missmat")  # misclassification matrix
plot_vacalib(vacalib_fit = vacalib_out_eava, toplot = "csmf")  # CSMFs
plot_vacalib(vacalib_fit = vacalib_out_eava, toplot = "both")  # both


# InSilicoVA
vacalib_out_insilicova = vacalibration(va_data = comsamoz_CCVAoutput$neonate[2],
                                       age_group = "neonate", country = "Mozambique",
                                       saveoutput = FALSE)

# summary plot
plot_vacalib(vacalib_fit = vacalib_out_insilicova, toplot = "missmat")  # misclassification matrix
plot_vacalib(vacalib_fit = vacalib_out_insilicova, toplot = "csmf")  # CSMFs
plot_vacalib(vacalib_fit = vacalib_out_insilicova, toplot = "both")  # both


# InterVA
vacalib_out_interva = vacalibration(va_data = comsamoz_CCVAoutput$neonate[3],
                                    age_group = "neonate", country = "Mozambique",
                                    saveoutput = FALSE)

# summary plot
plot_vacalib(vacalib_fit = vacalib_out_interva, toplot = "missmat")  # misclassification matrix
plot_vacalib(vacalib_fit = vacalib_out_interva, toplot = "csmf")  # CSMFs
plot_vacalib(vacalib_fit = vacalib_out_interva, toplot = "both")  # both



######### Ensemble Calibration #########
vacalib_out_ensemble = vacalibration(va_data = comsamoz_CCVAoutput$neonate,
                                     age_group = "neonate", country = "Mozambique",
                                     saveoutput = FALSE)

# summary plot
plot_vacalib(vacalib_fit = vacalib_out_ensemble, toplot = "missmat")  # misclassification matrix
plot_vacalib(vacalib_fit = vacalib_out_ensemble, toplot = "csmf")  # CSMFs
plot_vacalib(vacalib_fit = vacalib_out_ensemble, toplot = "both")  # both
}

}
