% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vacalib_fixed.R
\name{plot_vacalib_fixed}
\alias{plot_vacalib_fixed}
\title{Summary Plots of VA-Calibration Using Fixed Misclassification Matrix}
\usage{
plot_vacalib_fixed(vacalib_fit, toplot)
}
\arguments{
\item{vacalib_fit}{Fitted object from \code{vacalibration()}}

\item{toplot}{Character. Same as \code{toplot} in \code{plot_vacalib_fixed()}}
}
\value{
Plots misclassification matrices and/or cause-specific mortality fractions
}
\description{
This is a utility function. Please use \link{plot_vacalib}.
}
