% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_round.R
\name{smart_round}
\alias{smart_round}
\title{Round and maintain a target sum}
\usage{
smart_round(x, target_sum, digits = 0)
}
\arguments{
\item{x}{Numeric vector.}

\item{target_sum}{Numeric. The target sum to be maintained after rounding. Default is \code{NULL} which sets \code{target_sum=sum(x)}.}

\item{digits}{Positive integer. Indicates the number of decimal places to be used.}
}
\value{
Numeric vector.
}
\description{
Rounds a vector to the specified number of decimal places and maintains the sum it had before rounding.
}
\examples{

x = rep(1/3, 3)
round(x, 2)
smart_round(x, 1, 2)

}
