% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacalibration.R
\name{vacalibration}
\alias{vacalibration}
\title{VA-Calibration}
\usage{
vacalibration(
  va_data = NULL,
  age_group = NULL,
  country = NULL,
  missmat_type = c("prior", "fixed", "samples")[1],
  studycause_map = NULL,
  missmat = NULL,
  donotcalib = NULL,
  donotcalib_type = c("learn", "fixed")[1],
  nocalib.threshold = 0.1,
  path_correction = TRUE,
  ensemble = NULL,
  pshrink_strength = NULL,
  nMCMC = 5000,
  nBurn = 5000,
  nThin = 1,
  nChain = 1,
  nCore = 1,
  adapt_delta_stan = 0.9,
  refresh_stan = NULL,
  seed = 1,
  verbose = TRUE,
  saveoutput = FALSE,
  output_filename = NULL,
  output_dir = NULL
)
}
\arguments{
\item{va_data}{Named list. Algorithm-specific unlabeled VA data.

It expects a named list, such as \code{list("algo1" = algo1_output, "algo2" = algo2_output, ...)}.

Misclassification matrix estimates in \code{CCVA_missmat} are only available for CCVA algorithms \href{https://www.jogh.org/documents/issue201601/jogh-06-010601.pdf}{EAVA}, \href{https://www.tandfonline.com/doi/abs/10.1080/01621459.2016.1152191}{InSilicoVA}, and InterVA (\doi{10.3402/gha.v5i0.19281}). For them the algorithm names in input data must be \code{"eava"}, \code{"insilicova"}, and \code{"interva"}. Otherwise, users must input misclassification matrices in \code{missmat} (see more details in \code{missmat}).

VA data provided for each algorithm (\code{algo1_output}, \code{algo2_output}, ...) can be either
\enumerate{
\item outputs of CCVA algorithms (output from \code{codEAVA()} in \code{EAVA} for EAVA, and \code{codeVA()} and \code{prepCalibration()} in \code{openVA} for InSilicoVA and InterVA), or
\item individual broad cause of deaths (output from \code{cause_map}), or
\item a vector of cause-specific death counts.
}

More generally, it can calibrate for any discrete classifier. In that case, the input must be one of these two types:
\enumerate{
\item A binary matrix arranged as individuals along rows and class labels as columns. For each individual (row), 1 occurs exactly once and it indicates the estimated class label. Other elements in the row are 0.
\item A vector of label-specific counts. This indicates the estimated number of individuals for each label.
}}

\item{age_group}{Character.

When \code{missmat} is \code{NULL}, this indicates the age group for which the misclassification matrix estimates in \code{"CCVA_missmat"} should be applied (default).

It can be either \code{"neonate"} for neonatal deaths occurring between 0-27 days after birth, and \code{"child"} for deaths among children occurring between 1-59 months.}

\item{country}{Character.

When \code{missmat} is \code{NULL}, this indicates the country for which the misclassification matrix estimates in \code{"CCVA_missmat"} should be applied (default).

If input is "Bangladesh", "Ethiopia", "Kenya", "Mali", "Mozambique", "Sierra Leone", or "South Africa", then their corresponding misclassification matrix is applied. For any other country, the estimate for "other" is applied (see \code{"CCVA_missmat"} for more details).}

\item{missmat_type}{Character. Indicates the type of misclassification matrix estimates provided in \code{missmat}.

\code{"prior"} (default) Dirichlet priors for each row of the misclassification matrix.

\code{"fixed"} A fixed misclassification matrix.

\code{"samples"} Random samples of misclassification matrix.

Uncertainty in misclassification matrix estimates is only propagated for \code{"prior"} or \code{"samples"}.}

\item{studycause_map}{Named character vector. A mapping of observed causes (in \code{va_data}) to broad causes (for which misclassification estimates are available in \code{"CCVA_missmat"}).

Required only when \code{missmat} is \code{NULL}, and causes observed in \code{va_data} are not a subset of broad causes in \code{"CCVA_missmat"} (see \code{"CCVA_missmat"} for list of causes).

For example, if causes observed in \code{va_data} for neonates are "cause1", "cause2", "cause3", and "cause4", \code{studycause_map} expects input as \code{c("cause1" = "pneumonia", "cause2" = "ipre", "cause3" = "other", "cause4" = "other")}.}

\item{missmat}{Named list. Similarly structured as \code{va_data}. For example, \code{list("algo1" = missmat_algo1, "algo2" = missmat_algo2, ...)}.

For \code{missmat_type = "prior"}, \code{missmat_algo1}, \code{missmat_algo2}, ... are matrices with positive entries and arranged as CHAMPS cause \out{×} VA cause. Each row of the matrix is a vector of Dirichlet scale parameters. This the Dirichlet prior assumed on the corresponding row of the misclassification matrix. See stored estimates \code{CCVA_missmat$neonate$eava$asDirich$Mozambique} for example.

For \code{missmat_type = "fixed"}, \code{missmat_algo1}, \code{missmat_algo2}, ... are misclassification matrices arranged as CHAMPS cause \out{×} VA cause. See stored estimates \code{CCVA_missmat$neonate$eava$postmean$Mozambique} for example.

For \code{missmat_type = "samples"}, \code{missmat_algo1}, \code{missmat_algo2}, ... are arrays of misclassification matrix samples arranged as samples \out{×} CHAMPS cause \out{×} VA cause. \code{missmat_algo1[i,,]} is the i-th sample of misclassification matrix for \code{algo1}. See the samples stored in the \href{https://github.com/sandy-pramanik/CCVA-Misclassification-Matrices}{CCVA-Misclassification-Matrices} GitHub repository for example.

Names and length of \code{missmat} must be identical to \code{va_data}.

Users are not required to provide \code{missmat} for using the stored estimates in \code{"CCVA_missmat"}. They can simply input the required \code{age_group}, \code{country}, \code{missmat_type}, and \code{studycause_map} accordingly.

\code{missmat} needs to be input when causes observed in \code{va_data} are not a subset of CHAMPS broad causes (in \code{"CCVA_missmat"}) and \code{studycause_map} is not provided.

For a general purpose of calibrating categorical classifiers, CHAMPS and VA causes can be interpreted as true and estimated labels and users must input \code{missmat}.}

\item{donotcalib}{Named list. List of causes for each algorithm that users do not want to calibrate. The set of causes can differ across algorithms.

Default: \code{list("eava"="other", "insilicova"="other", "interva"="other")}. When using the stored estimates in \code{CCVA_missmat}, this implies that the cause-specific mortality fractions (CSMF) for CHAMPS broad cause \code{"other"} is not calibrated.

When causes observed in \code{va_data} are not a subset of CHAMPS broad causes and \code{studycause_map} is provided, all observed causes in \code{va_data} that match with the causes in \code{donotcalib} are not calibrated.

Set \code{list("eava"=NULL, "insilicova"=NULL, "interva"=NULL)} to calibrate all causes.

For a general purpose of calibrating categorical classifiers, causes can be interpreted as class labels and specified accordingly.}

\item{donotcalib_type}{Character. \code{"learn"} (default) or \code{"fixed"}.

For \code{donotcalib_type="fixed"}, only the causes specified in \code{"donotcalib"} are not calibrated.

For \code{donotcalib_type="learn"}, it learns additional causes from misclassification matrix in \code{"missmat"} that cannot be calibrated.

When misclassification rates for a VA cause do not change across CHAMPS causes, the calibration equation becomes underdetermined (see the footnote on pg. 1227 in \href{https://projecteuclid.org/journals/annals-of-applied-statistics/volume-19/issue-2/Modeling-structure-and-country-specific-heterogeneity-in-misclassification-matrices-of/10.1214/24-AOAS2006.full}{Pramanik et al. (2025)}). When \code{donotcalib_type="learn"}, it screens VA causes that do not vary beyond \code{nocalib.threshold}. These causes are added to the \code{donotcalib} list.

For a general purpose of calibrating categorical classifiers, causes can be interpreted as class labels and specified accordingly.}

\item{nocalib.threshold}{Numeric in \code{(0,1)}.

The threshold used to screen VA causes when \code{donotcalib_type="learn"}.

Default: 0.1.}

\item{path_correction}{Logical. Setting \code{TRUE} shrinks misclassification matrix towards the identity matrix to improve stability in VA-Calibration.

Default is \code{TRUE}.}

\item{ensemble}{Logical. Whether to perform ensemble calibration when outputs from multiple algorithms are provided.

Default is \code{TRUE}.}

\item{pshrink_strength}{Positive numeric. Degree of shrinkage of calibrated CSMF estimates towards its uncalibrated estimates. This is the parameter \code{eta} in the prior of calibrated CSMF \code{p} (see pg. 1226 in \href{https://projecteuclid.org/journals/annals-of-applied-statistics/volume-19/issue-2/Modeling-structure-and-country-specific-heterogeneity-in-misclassification-matrices-of/10.1214/24-AOAS2006.full}{Pramanik et al. (2025)}).

Only used when \code{path_correction=FALSE}. \code{pshrink_strength} is set to 0 when \code{path_correction=TRUE}.

Defaults to 4 when \code{path_correction=FALSE}.}

\item{nMCMC}{Positive integer. Total number of posterior samples to perform inference on.

Total number of iterations are \code{nBurn + nMCMC*nThin}. Default 5000.}

\item{nBurn}{Positive integer. Total burn-in in posterior sampling.

Total number of iterations are \code{nBurn + nMCMC*nThin}. Default 5000.}

\item{nThin}{Positive integer. Number of thinning in posterior sampling.

Total number of iterations are \code{nBurn + nMCMC*nThin}. Default 1.}

\item{nChain}{Positive integer. Number of chains for Stan sampling. Default 1.}

\item{nCore}{Positive integer. Number of cores to run multiple chains in parallel for Stan sampling. Default 1.}

\item{adapt_delta_stan}{Numeric in \code{(0,1)}. \code{adapt_delta} parameter in \code{rstan}.

Influences the behavior of the No-U-Turn Sampler (NUTS) in Stan.

Default 0.9.}

\item{refresh_stan}{Positive integer. Print every \code{refresh_stan}\% progress.

Default 20.}

\item{seed}{Numeric. \code{seed} parameter in rstan. Default 1.}

\item{verbose}{Logical. Whether to report progress (\code{TRUE}) or not (\code{FALSE}).

Default \code{TRUE}.}

\item{saveoutput}{Logical. Save output (\code{TRUE}) or not (\code{FALSE}).

Default \code{TRUE}.}

\item{output_filename}{Character. Output name to save as.

Default \code{vacalibration_out}.}

\item{output_dir}{Output directory or file path to save at.

Default \code{getwd()}, the working directory.}
}
\value{
A list with components:
\itemize{
\item \code{calib_MCMCout} — Output from Stan fits.
\item \code{p_uncalib} — Uncalibrated estimates of CSMF. It is a matrix arranged as algorithm \out{×} VA causes (estimated labels).
\item \code{p_calib} — Posterior samples of calibrated CSMF. It is an array arranged as algorithm \out{×} samples \out{×} VA causes (or estimated labels).
\item \code{pcalib_postsumm} — Posterior summaries (mean and 95\% credible interval) of calibrated CSMF. It is an array arranged as algorithm \out{×} summary measures \out{×} VA causes (or estimated labels).
\item \code{va_deaths_uncalib} — Uncalibrated cause-specific death counts. It is a matrix arranged as algorithm \out{×} VA causes (or estimated labels).
\item \code{va_deaths_calib_algo} — Calibrated cause-specific death counts from algorithm-specific calibration. It is a matrix arranged as algorithm \out{×} VA causes (or estimated labels).
\item \code{va_deaths_calib_ensemble} — Calibrated cause-specific death counts from ensemble calibration. It is a matrix arranged as algorithm \out{×} VA causes (or estimated labels).
\item \code{Mmat_input} — \code{"missmat"} as provided in the input. It is an array arranged as algorithm \out{×} CHAMPS cause (or true labels) \out{×} VA causes (or estimated labels).
\item \code{Mmat_study} — Modified \code{Mmat_input} if \code{studycause_map} is provided. It is an array arranged in the same way as \code{Mmat_input}.
\item \code{Mmat_tomodel} — Modified \code{Mmat_study} if \code{path_correction} is \code{TRUE}. This is used for calibration. It is an array arranged in the same way as \code{Mmat_input} and \code{Mmat_study}.
\item \code{donotcalib_study} — This indicates causes that are not calibrated for each algorithm, as specified in the input \code{donotcalib}. It is a logical matrix arranged as algorithm \out{×} VA causes (or estimated labels).
\item \code{donotcalib_tomodel} — This indicates causes that are not calibrated in each calibration. This is a modified \code{donotcalib_study} if \code{donotcalib_type} is provided and \code{ensemble=TRUE}. It is a logical matrix arranged as algorithm \out{×} VA causes (or estimated labels).
\item \code{calibrated} — \code{TRUE} or \code{FALSE} indicating whether Stan sampling was performed for calibration.
\item \code{lambda_calibpath} — When \code{path_correction=TRUE}, this indicates the degree of shrinkage of CSMF for each algorithm towards uncalibrated estimates. This is a vector of numerics in \code{[0,1]} showing degrees of shrinkage for each algorithm.
\item \code{K} — Number of algorithms.
\item \code{nCause} — Number of causes.
\item \code{causes} — Name of causes.
\item \code{input} — List of inputs.
}
}
\description{
This is the main function in the package. It calibrates population-level cause-specific mortality fractions (CSMFs) that are derived using computer-coded verbal autopsy (CCVA) algorithms. For VA-Calibration, the function utilizes the inventory of misclassification matrix estimates \link{CCVA_missmat}.
The outputs from \href{https://CRAN.R-project.org/package=EAVA}{EAVA} and \href{https://CRAN.R-project.org/package=openVA}{openVA} for InSilicoVA and InterVA can be input directly (see below). This seamlessly supports VA-Calibration for \href{https://www.jogh.org/documents/issue201601/jogh-06-010601.pdf}{EAVA}, \href{https://www.tandfonline.com/doi/abs/10.1080/01621459.2016.1152191}{InSilicoVA}, and InterVA (\doi{10.3402/gha.v5i0.19281}).
For other CCVA algorithms, the input expects either an individual by cause matrix, or cause-specific death count vector (see below).
When broad-cause-specific death counts are input and they do not match the broad causes in the stored misclassification estimates, then either \code{studycause_map} or the misclassification matrices (fixed or as row-specific Dirichlet priors) need to be provided.
More generally, this allows us to calibrate population-level prevalence derived from single-class predictions of discrete classifiers. For this, users need to provide fixed or uncertainty-quantified misclassification matrices.
}
\examples{

\donttest{
######### COMSA-Mozambique VA-COD data #########
data(comsamoz_CCVAoutput)

# neonatal deaths
comsamoz_CCVAoutput$neonate$eava  # output from running EAVA
comsamoz_CCVAoutput$neonate$insilicova  # output from running InSilicoVA
comsamoz_CCVAoutput$neonate$interva  # output from running InterVA



######### Algorithm-Specific Calibration #########

# EAVA
vacalib_out_eava = vacalibration(va_data = comsamoz_CCVAoutput$neonate[1],
                                 age_group = "neonate", country = "Mozambique",
                                 saveoutput = FALSE)

## CSMF
vacalib_out_eava$p_uncalib   # uncalibrated
vacalib_out_eava$p_calib   # calibrated
vacalib_out_eava$pcalib_postsumm   # summary of calibrated estimates

## death counts
vacalib_out_eava$va_deaths_uncalib   # uncalibrated
vacalib_out_eava$va_deaths_calib_algo   # calibrated


# InSilicoVA
vacalib_out_insilicova = vacalibration(va_data = comsamoz_CCVAoutput$neonate[2],
                                       age_group = "neonate", country = "Mozambique",
                                       saveoutput = FALSE)

## CSMF
vacalib_out_insilicova$p_uncalib   # uncalibrated
vacalib_out_insilicova$p_calib   # calibrated
vacalib_out_insilicova$pcalib_postsumm   # summary of calibrated estimates

## death counts
vacalib_out_insilicova$va_deaths_uncalib   # uncalibrated
vacalib_out_insilicova$va_deaths_calib_algo   # calibrated


# InterVA
vacalib_out_interva = vacalibration(va_data = comsamoz_CCVAoutput$neonate[3],
                                    age_group = "neonate", country = "Mozambique",
                                    saveoutput = FALSE)

## CSMF
vacalib_out_interva$p_uncalib   # uncalibrated
vacalib_out_interva$p_calib   # calibrated
vacalib_out_interva$pcalib_postsumm   # summary of calibrated estimates

## death counts
vacalib_out_interva$va_deaths_uncalib   # uncalibrated
vacalib_out_interva$va_deaths_calib_algo   # calibrated



######### Ensemble Calibration #########
vacalib_out_ensemble = vacalibration(va_data = comsamoz_CCVAoutput$neonate,
                                     age_group = "neonate", country = "Mozambique",
                                     saveoutput = FALSE)

## CSMF
vacalib_out_ensemble$p_uncalib   # uncalibrated
vacalib_out_ensemble$p_calib   # calibrated
vacalib_out_ensemble$pcalib_postsumm   # summary of calibrated estimates

## death counts
vacalib_out_ensemble$va_deaths_uncalib   # uncalibrated
vacalib_out_ensemble$va_deaths_calib_algo   # algorithm-specific calibrated death counts
vacalib_out_ensemble$va_deaths_calib_ensemble   # ensemble calibrated death counts
}

}
\references{
Pramanik, S, et al. (2026)
Country-Specific Estimates of Misclassification Rates of Computer-Coded Verbal Autopsy Algorithms
BMJ Global Health
\doi{10.1136/bmjgh-2025-021747}

Pramanik, S, et al. (2025)
Modeling structure and country-specific heterogeneity in misclassification matrices of verbal autopsy-based cause of death classifiers
Annals of Applied Statistics
\href{https://projecteuclid.org/journals/annals-of-applied-statistics/volume-19/issue-2/Modeling-structure-and-country-specific-heterogeneity-in-misclassification-matrices-of/10.1214/24-AOAS2006.short}{Link}

Fiksel, J., et al. (2022)
Generalized Bayes Quantification Learning under Dataset Shift
Journal of the American Statistical Association
\href{https://www.tandfonline.com/doi/full/10.1080/01621459.2021.1909599}{Link}

Datta, A, et al. (2021)
Regularized Bayesian transfer learning for population-level etiological distributions.
Biostatistics
\doi{10.1093/biostatistics/kxaa001}
}
