% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_vacalib_prior.R
\name{modular_vacalib_prior}
\alias{modular_vacalib_prior}
\title{Modular VA-Calibration for Dirichlet Prior on Misclassification Matrix}
\usage{
modular_vacalib_prior(
  va_unlabeled = NULL,
  Mmat_calib = NULL,
  studycause_map = NULL,
  donotcalib = NULL,
  donotcalib_type = c("learn", "fixed")[1],
  nocalib.threshold = 0.1,
  path_correction = TRUE,
  ensemble = NULL,
  shrink_strength = NULL,
  lambda_eBayes = NULL,
  nMCMC = 5000,
  nBurn = 5000,
  nThin = 1,
  nChain = 1,
  nCore = 1,
  adapt_delta_stan = 0.9,
  refresh.stan = NULL,
  seed = 1,
  verbose = TRUE,
  plot_it = TRUE
)
}
\description{
Modular VA-Calibration for Dirichlet Prior on Misclassification Matrix
}
